add tests for edit functionality
