+++ package.json
@@ -44,5 +44,8 @@
     "gh-pages": "^5.0.0",
     "lodash": "^4.17.21"
   },
-  "main": "index.js"
+  "main": "index.js",
+  "jest": {
+    "testEnvironment": "jsdom"
+  }
 }
